refactor(services): split cell_methods/inventory.rs into submodules
Separated method indices, dispatch, item ops, bandolier handlers, weapon events, and tests into focused files mirroring the cell_methods/player/ pattern.

Public API unchanged. No behavior changes.
